requirements.txt
# Core & System
python-dotenv>=1.2
requests==2.33.1
psutil==7.2.2
colorama==0.4.6
pydantic-settings==2.13.1
PyYAML>=6.0.1

# Parsing Engine
curl_cffi>=0.15
beautifulsoup4>=4.14
jmespath>=1.1
jsonschema>=4.26

# Browser
camoufox[geoip]>=0.4.11

# Data & Exports
openpyxl>=3.1.5

# UI
flet>=0.84.0

# Dev / Build
rich
Pillow>=11.0.0

# Tests & Formatters
pytest==9.0.2
pytest-mock==3.15.1
pytest-cov>=7.1.0
pytest-asyncio>=1.3
pytest-httpserver>=1.1.5